+++ CHANGELOG.md
@@ -1,7 +1,26 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+
 ## [Unreleased]
+- Planned User registeration.
+- Planned authorization.
+
+
+
+## [1.1.0]
+### Added 
+- added views for listing items, expired items, adding items, removing/consuming items
+- Implemented Authtication via JWT to views
+- added migrations 
+- Handled inventory logic: add to existing item if expiry matches, create a new item if expiry differs
+- Prevented the use of expired items in the inventory.
+
+
+
+
+
+## [1.0.0]
 ### Added
 - Initial changelog setup.
 - Added apps Inventory and Kitchen